Rename file. Add java preloader file.
