+++ rayvens/cli/java.py
@@ -1,0 +1,50 @@
+#
+# Copyright IBM Corporation 2021
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+
+import rayvens.cli.file as file
+
+# The preloader file is fixed and can just be a string.
+preloader_file_name = "Preloader.java"
+preloader_file_contents = """
+import org.apache.camel.BindToRegistry;
+import org.apache.camel.Exchange;
+import org.apache.camel.Processor;
+import org.apache.camel.builder.RouteBuilder;
+
+class Exit implements Processor {
+  public void process(Exchange exchange) throws Exception {
+    System.exit(0);
+  }
+}
+
+public class Preloader extends RouteBuilder {
+  @Override
+  public void configure() throws Exception {
+    from("timer:tick").to("bean:exit");
+    from("platform-http:/null").to("http:null");
+  }
+
+  @BindToRegistry
+  public Exit exit() {
+    return new Exit();
+  }
+}
+"""
+
+
+def write_preloader_file(directory):
+    preloader_file_path = directory.joinpath(preloader_file_name)
+    file.write_file(preloader_file_path, preloader_file_contents)
